refactor(metrics): move common logic to MetricManagaer base class
