+++ aws_lambda_powertools/metrics/metrics.py
@@ -1,15 +1,6 @@
-import functools
-import json
-import logging
-import warnings
-from typing import Any, Callable, Dict, Optional, Union
+from typing import Any, Dict
 
-from .base import MetricManager, MetricUnit
-from .metric import single_metric
-
-logger = logging.getLogger(__name__)
-
-is_cold_start = True
+from .base import MetricManager
 
 
 class Metrics(MetricManager):
@@ -55,9 +46,6 @@ class Metrics(MetricManager):
         service name to be used as metric dimension, by default "service_undefined"
     namespace : str, optional
         Namespace for metrics
-    singleton : boolean, by default True
-        Whether to reuse metrics data across Metrics instances (isolation)
-
 
     Raises
     ------
@@ -71,7 +59,7 @@ class Metrics(MetricManager):
 
     # NOTE: We use class attrs to share metrics data across instances
     # this allows customers to initialize Metrics() throughout their code base (and middlewares)
-    # and not get caught by accident with metrics data loss
+    # and not get caught by accident with metrics data loss, or data deduplication
     # e.g., m1 and m2 add metric ProductCreated, however m1 has 'version' dimension  but m2 doesn't
     # Result: ProductCreated is created twice as we now have 2 different EMF blobs
     _metrics: Dict[str, Any] = {}
@@ -79,39 +67,23 @@ class Metrics(MetricManager):
     _metadata: Dict[str, Any] = {}
     _default_dimensions: Dict[str, Any] = {}
 
-    def __init__(self, service: Optional[str] = None, namespace: Optional[str] = None, singleton: bool = True):
+    def __init__(self, service: str = "", namespace: str = ""):
         self.metric_set = self._metrics
         self.service = service
-        self.namespace: Optional[str] = namespace
+        self.namespace = namespace
         self.metadata_set = self._metadata
         self.default_dimensions = self._default_dimensions
         self.dimension_set = self._dimensions
 
-        # We couldn't find a better name for a new class; 'singleton' param fits the purpose
-        # Customers can now disable data sharing with singleton=False
-        # It unlocks distinct namespace metrics, multi-EMF blobs and multi-tenant use cases
-        # See https://github.com/awslabs/aws-lambda-powertools-python/issues/1668
-        if singleton:
-            self.dimension_set.update(**self._default_dimensions)
-            return super().__init__(
-                metric_set=self.metric_set,
-                dimension_set=self.dimension_set,
-                namespace=self.namespace,
-                metadata_set=self.metadata_set,
-                service=self.service,
-                default_dimensions=self.default_dimensions,
-            )
-
-        # NOTE: With class initialized, we can safely clean this instance attrs
-        # this ensures instantiating a non-singleton instance after a singleton instance
-        # won't affect metric data sets, as we have different memory pointers
-        # e.g., `self._metrics != Metrics._metrics` (instance attrs vs class attrs)
-        self._metrics = {}
-        self._dimensions = {}
-        self._metadata = {}
-        self._default_dimensions = {}
-
-        return super().__init__()
+        self.dimension_set.update(**self._default_dimensions)
+        return super().__init__(
+            metric_set=self.metric_set,
+            dimension_set=self.dimension_set,
+            namespace=self.namespace,
+            metadata_set=self.metadata_set,
+            service=self.service,
+            default_dimensions=self.default_dimensions,
+        )
 
     def set_default_dimensions(self, **dimensions) -> None:
         """Persist dimensions across Lambda invocations
@@ -132,7 +104,7 @@ class Metrics(MetricManager):
 
             @metrics.log_metrics()
             def lambda_handler():
-                   return True
+                return True
         """
         for name, value in dimensions.items():
             self.add_dimension(name, value)
@@ -143,98 +115,6 @@ class Metrics(MetricManager):
         self.default_dimensions.clear()
 
     def clear_metrics(self) -> None:
-        logger.debug("Clearing out existing metric set from memory")
-        self.metric_set.clear()
-        self.dimension_set.clear()
-        self.metadata_set.clear()
-        self.set_default_dimensions(**self.default_dimensions)  # re-add default dimensions
-
-    def log_metrics(
-        self,
-        lambda_handler: Union[Callable[[Dict, Any], Any], Optional[Callable[[Dict, Any, Optional[Dict]], Any]]] = None,
-        capture_cold_start_metric: bool = False,
-        raise_on_empty_metrics: bool = False,
-        default_dimensions: Optional[Dict[str, str]] = None,
-    ):
-        """Decorator to serialize and publish metrics at the end of a function execution.
-
-        Be aware that the log_metrics **does call* the decorated function (e.g. lambda_handler).
-
-        Example
-        -------
-        **Lambda function using tracer and metrics decorators**
-
-            from aws_lambda_powertools import Metrics, Tracer
-
-            metrics = Metrics(service="payment")
-            tracer = Tracer(service="payment")
-
-            @tracer.capture_lambda_handler
-            @metrics.log_metrics
-            def handler(event, context):
-                    ...
-
-        Parameters
-        ----------
-        lambda_handler : Callable[[Any, Any], Any], optional
-            lambda function handler, by default None
-        capture_cold_start_metric : bool, optional
-            captures cold start metric, by default False
-        raise_on_empty_metrics : bool, optional
-            raise exception if no metrics are emitted, by default False
-        default_dimensions: Dict[str, str], optional
-            metric dimensions as key=value that will always be present
-
-        Raises
-        ------
-        e
-            Propagate error received
-        """
-
-        # If handler is None we've been called with parameters
-        # Return a partial function with args filled
-        if lambda_handler is None:
-            logger.debug("Decorator called with parameters")
-            return functools.partial(
-                self.log_metrics,
-                capture_cold_start_metric=capture_cold_start_metric,
-                raise_on_empty_metrics=raise_on_empty_metrics,
-                default_dimensions=default_dimensions,
-            )
-
-        @functools.wraps(lambda_handler)
-        def decorate(event, context):
-            try:
-                if default_dimensions:
-                    self.set_default_dimensions(**default_dimensions)
-                response = lambda_handler(event, context)
-                if capture_cold_start_metric:
-                    self.__add_cold_start_metric(context=context)
-            finally:
-                if not raise_on_empty_metrics and not self.metric_set:
-                    warnings.warn("No metrics to publish, skipping")
-                else:
-                    metrics = self.serialize_metric_set()
-                    self.clear_metrics()
-                    print(json.dumps(metrics, separators=(",", ":")))
-
-            return response
-
-        return decorate
-
-    def __add_cold_start_metric(self, context: Any) -> None:
-        """Add cold start metric and function_name dimension
-
-        Parameters
-        ----------
-        context : Any
-            Lambda context
-        """
-        global is_cold_start
-        if is_cold_start:
-            logger.debug("Adding cold start metric and function_name dimension")
-            with single_metric(name="ColdStart", unit=MetricUnit.Count, value=1, namespace=self.namespace) as metric:
-                metric.add_dimension(name="function_name", value=context.function_name)
-                if self.service:
-                    metric.add_dimension(name="service", value=str(self.service))
-                is_cold_start = False
+        super().clear_metrics()
+        # re-add default dimensions
+        self.set_default_dimensions(**self.default_dimensions)
